Modified it to dynamically load servers from AWS EC2's.
This change isn't the most pretty thing you will see, but it gets the
job done.

Allowing site name and a site logo to be set in settings.

Removed the 'authenticated' links to individual Supervisor instances
since they contained the username and password in clear text as part
of the URL! WTF were you thinking!? If you want to access the servers
directly, you will need to authenticate manually now.

Other improvements, now that EC2 instances are used:
- Ability to filter the servers based on the EC2 Tag 'Name'. Default
  filter buttons are setup for alpha, beta, and production.
- Added another row to the server table to show some other Tags from
  the EC2 instances, as well as their public and private IP addresses.